// apps/web/src/api/generated/models/embed/embed.zod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Talqo API
 * OpenAPI spec version: 0.0.1
 */
import * as zod from "zod"

import { EmbedAppearance } from "./embedAppearance.zod"

export const embedAccessVersionExclusiveMin = 0

export const Embed = zod.object({
	id: zod.string(),
	agentId: zod.string(),
	name: zod.string(),
	embedToken: zod.string(),
	accessVersion: zod.int().gt(embedAccessVersionExclusiveMin),
	appearance: EmbedAppearance,
})

export type Embed = zod.input<typeof Embed>
export type EmbedOutput = zod.output<typeof Embed>